Cast strtable_get_str() to uintptr_t

+++ src/module.c
@@ -53,7 +53,7 @@ static uintptr_t dc_event_handler(dc_context_t* dc_context, int event, uintptr_t
       return dcn_context->is_offline;
 
     case DC_EVENT_GET_STRING:
-      return strtable_get_str(dcn_context->strtable, (int)data1);
+      return (uintptr_t)strtable_get_str(dcn_context->strtable, (int)data1);
 
     case DC_EVENT_HTTP_GET: {
       uintptr_t http_ret = 0;
